Fix windows

+++ fmod-oxide/src/core/enums.rs
@@ -12,7 +12,9 @@ use fmod_sys::*;
     num_enum::IntoPrimitive,
     num_enum::UnsafeFromPrimitive
 )]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum SpeakerMode {
     Default = FMOD_SPEAKERMODE_DEFAULT,
     Raw = FMOD_SPEAKERMODE_RAW,
@@ -31,7 +33,9 @@ pub enum SpeakerMode {
     num_enum::IntoPrimitive,
     num_enum::UnsafeFromPrimitive
 )]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum OutputType {
     AutoDetect = FMOD_OUTPUTTYPE_AUTODETECT,
     Unknown = FMOD_OUTPUTTYPE_UNKNOWN,
@@ -62,7 +66,9 @@ pub enum OutputType {
     num_enum::IntoPrimitive,
     num_enum::UnsafeFromPrimitive
 )]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum ThreadType {
     Mixer = FMOD_THREAD_TYPE_MIXER,
     Feeder = FMOD_THREAD_TYPE_FEEDER,
@@ -126,7 +132,9 @@ pub enum Speaker {
     num_enum::IntoPrimitive,
     num_enum::UnsafeFromPrimitive
 )]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum PluginType {
     Output = FMOD_PLUGINTYPE_OUTPUT,
     Codec = FMOD_PLUGINTYPE_CODEC,
@@ -139,7 +147,9 @@ pub enum PluginType {
     num_enum::IntoPrimitive,
     num_enum::UnsafeFromPrimitive
 )]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum DspType {
     Unknown = FMOD_DSP_TYPE_UNKNOWN,
     Mixer = FMOD_DSP_TYPE_MIXER,
@@ -186,7 +196,9 @@ pub enum DspType {
     num_enum::IntoPrimitive,
     num_enum::UnsafeFromPrimitive
 )]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum PortType {
     Music = FMOD_PORT_TYPE_MUSIC,
     CopyrightMusic = FMOD_PORT_TYPE_COPYRIGHT_MUSIC,
@@ -203,7 +215,9 @@ pub enum PortType {
     num_enum::IntoPrimitive,
     num_enum::UnsafeFromPrimitive
 )]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum SoundGroupBehavior {
     Fail = FMOD_SOUNDGROUP_BEHAVIOR_FAIL,
     Mute = FMOD_SOUNDGROUP_BEHAVIOR_MUTE,
@@ -216,7 +230,9 @@ pub enum SoundGroupBehavior {
     num_enum::IntoPrimitive,
     num_enum::UnsafeFromPrimitive
 )]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum DspConnectionType {
     Standard = FMOD_DSPCONNECTION_TYPE_STANDARD,
     Sidechain = FMOD_DSPCONNECTION_TYPE_SIDECHAIN,
@@ -243,7 +259,9 @@ pub enum DspParameterDataType {
     num_enum::IntoPrimitive,
     num_enum::UnsafeFromPrimitive
 )]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum SoundType {
     Unknown = FMOD_SOUND_TYPE_UNKNOWN,
     AIFF = FMOD_SOUND_TYPE_AIFF,
@@ -278,7 +296,9 @@ pub enum SoundType {
     num_enum::IntoPrimitive,
     num_enum::UnsafeFromPrimitive
 )]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum SoundFormat {
     None = FMOD_SOUND_FORMAT_NONE,
     PCM8 = FMOD_SOUND_FORMAT_PCM8,
@@ -295,7 +315,9 @@ pub enum SoundFormat {
     num_enum::IntoPrimitive,
     num_enum::UnsafeFromPrimitive
 )]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum TagType {
     Unknown = FMOD_TAGTYPE_UNKNOWN,
     ID3V1 = FMOD_TAGTYPE_ID3V1,
@@ -311,7 +333,9 @@ pub enum TagType {
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum OpenState {
     Ready = FMOD_OPENSTATE_READY,
     Loading = FMOD_OPENSTATE_LOADING,
@@ -338,7 +362,7 @@ impl OpenState {
             FMOD_OPENSTATE_SETPOSITION => Ok(OpenState::SetPosition),
             _ => Err(Error::EnumFromPrivitive {
                 name: "LoadingState",
-                primitive: value,
+                primitive: i64::from(value),
             }),
         }
     }
