+++ fmod-oxide/src/studio/enums.rs
@@ -9,7 +9,9 @@ use lanyard::Utf8CString;
 use std::ffi::{c_float, c_int};
 
 #[derive(Debug, Clone, PartialEq, Eq)]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum LoadingState {
     Unloading = FMOD_STUDIO_LOADING_STATE_UNLOADING,
     Unloaded = FMOD_STUDIO_LOADING_STATE_UNLOADED,
@@ -30,7 +32,7 @@ impl LoadingState {
                 .ok_or(Error::Fmod(FMOD_RESULT::FMOD_ERR_INVALID_PARAM)),
             _ => Err(Error::EnumFromPrivitive {
                 name: "LoadingState",
-                primitive: value,
+                primitive: i64::from(value),
             }),
         }
     }
@@ -42,7 +44,9 @@ impl LoadingState {
     num_enum::IntoPrimitive,
     num_enum::UnsafeFromPrimitive
 )]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum StopMode {
     AllowFadeout = FMOD_STUDIO_STOP_ALLOWFADEOUT,
     Immediate = FMOD_STUDIO_STOP_IMMEDIATE,
@@ -54,7 +58,9 @@ pub enum StopMode {
     num_enum::IntoPrimitive,
     num_enum::UnsafeFromPrimitive
 )]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum ParameterKind {
     GameControlled = FMOD_STUDIO_PARAMETER_GAME_CONTROLLED,
     AutomaticDistance = FMOD_STUDIO_PARAMETER_AUTOMATIC_DISTANCE,
@@ -82,7 +88,9 @@ pub enum UserPropertyKind {
     num_enum::IntoPrimitive,
     num_enum::UnsafeFromPrimitive
 )]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum PlaybackState {
     Playing = FMOD_STUDIO_PLAYBACK_PLAYING,
     Sustaining = FMOD_STUDIO_PLAYBACK_SUSTAINING,
@@ -97,7 +105,9 @@ pub enum PlaybackState {
     num_enum::IntoPrimitive,
     num_enum::UnsafeFromPrimitive
 )]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum EventProperty {
     ChannelPriority = FMOD_STUDIO_EVENT_PROPERTY_CHANNELPRIORITY,
     ScheduleDelay = FMOD_STUDIO_EVENT_PROPERTY_SCHEDULE_DELAY,
@@ -113,7 +123,9 @@ pub enum EventProperty {
     num_enum::IntoPrimitive,
     num_enum::UnsafeFromPrimitive
 )]
-#[repr(u32)]
+// stupid enum repr hack
+#[cfg_attr(target_os = "windows", repr(i32))]
+#[cfg_attr(target_os = "linux", repr(u32))]
 pub enum InstanceType {
     None = FMOD_STUDIO_INSTANCETYPE_NONE,
     System = FMOD_STUDIO_INSTANCETYPE_SYSTEM,
